TrabalhoRamiroeVitor2/requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.2.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
python-dotenv==1.0.1
bleach==6.2.0
PyJWT==2.9.0
Flask-SocketIO==5.3.6
eventlet==0.36.1
Flask-Limiter==3.8.0